Refactor product presentation dto
