+++ src/products/controllers/product-image.controller.ts
@@ -12,7 +12,6 @@ import {
   Post,
 } from '@nestjs/common';
 import { ApiOperation, ApiResponse, ApiBearerAuth } from '@nestjs/swagger';
-import { ProductsService } from '../products.service';
 import { ProductImage } from '../entities/product-image.entity';
 import { AuthGuard } from 'src/auth/auth.guard';
 import { Roles } from 'src/auth/roles.decorador';
@@ -22,10 +21,15 @@ import {
   CreateProductImageDTO,
   UpdateProductImageDTO,
 } from '../dto/product-image.dto';
+import { ProductImageService } from '../services/product-image.service';
+import { GenericProductService } from '../services/generic-product.service';
 
 @Controller('product/:productId/image')
 export class ProductImageController {
-  constructor(private readonly productsService: ProductsService) {}
+  constructor(
+    private readonly genericProductService: GenericProductService,
+    private readonly productImageService: ProductImageService,
+  ) {}
 
   @Get()
   @HttpCode(HttpStatus.OK)
@@ -34,11 +38,8 @@ export class ProductImageController {
   async getProductImages(
     @Param('productId') productId: string,
   ): Promise<ProductImage[]> {
-    const product = await this.productsService.findOne(productId);
-    if (!product) {
-      throw new NotFoundException('Product not found');
-    }
-    return product.images;
+    const images = await this.productImageService.findByProductId(productId);
+    return images;
   }
 
   @HttpCode(HttpStatus.CREATED)
@@ -52,11 +53,11 @@ export class ProductImageController {
     @Param('productId') productId: string,
     @Body() createProductImageDto: CreateProductImageDTO,
   ): Promise<void> {
-    const product = await this.productsService.findOne(productId);
+    const product = await this.genericProductService.findOne(productId);
     if (!product) {
       throw new NotFoundException('Product not found');
     }
-    await this.productsService.createProductImage(product, [
+    await this.productImageService.createProductImage(product, [
       createProductImageDto.url,
     ]);
   }
@@ -71,7 +72,7 @@ export class ProductImageController {
     @Param('productId') productId: string,
     @Param('imageId') imageId: string,
   ): Promise<ProductImage> {
-    const image = await this.productsService.findProductImage(
+    const image = await this.productImageService.findProductImage(
       productId,
       imageId,
     );
@@ -93,7 +94,7 @@ export class ProductImageController {
     @Param('imageId') imageId: string,
     @Body() updateProductImageDto: UpdateProductImageDTO,
   ): Promise<ProductImage> {
-    const image = await this.productsService.findProductImage(
+    const image = await this.productImageService.findProductImage(
       productId,
       imageId,
     );
@@ -101,7 +102,7 @@ export class ProductImageController {
       throw new NotFoundException('Product image not found');
     }
     image.url = updateProductImageDto.url ?? image.url;
-    return await this.productsService.updateProductImage(image);
+    return await this.productImageService.updateProductImage(image);
   }
 
   @Delete(':imageId')
@@ -116,13 +117,13 @@ export class ProductImageController {
     @Param('productId') productId: string,
     @Param('imageId') imageId: string,
   ): Promise<void> {
-    const image = await this.productsService.findProductImage(
+    const image = await this.productImageService.findProductImage(
       productId,
       imageId,
     );
     if (!image) {
       throw new NotFoundException('Product image not found');
     }
-    await this.productsService.deleteProductImage(image);
+    await this.productImageService.deleteProductImage(image);
   }
 }
