+++ src/products/products.controller.ts
@@ -1,32 +1,14 @@
-import {
-  Body,
-  Controller,
-  Get,
-  Post,
-  Query,
-  UseGuards,
-  UseInterceptors,
-} from '@nestjs/common';
+import { Controller, Get, Query, UseInterceptors } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import {
-  ApiBearerAuth,
-  ApiBody,
-  ApiCreatedResponse,
   ApiExtraModels,
   ApiOkResponse,
   ApiOperation,
   ApiQuery,
-  ApiUnauthorizedResponse,
   getSchemaPath,
 } from '@nestjs/swagger';
 import { ProductPresentationDTO, ProductQueryDTO } from './dto/product.dto';
 import { PaginationDTO } from 'src/utils/dto/pagination.dto';
-import { CreateProductDTO } from './dto/product.dto';
-import { Product } from './entities/product.entity';
-import { Roles } from 'src/auth/roles.decorador';
-import { AuthGuard } from 'src/auth/auth.guard';
-import { RolesGuard } from 'src/auth/roles.guard';
-import { UserRole } from 'src/user/entities/user.entity';
 import { PaginationInterceptor } from 'src/utils/pagination.interceptor';
 
 @Controller('product')
@@ -135,66 +117,5 @@ export class ProductsController {
     );
 
     return { data: products, total };
-  }
-
-  @Post()
-  @UseGuards(AuthGuard, RolesGuard)
-  @Roles(UserRole.ADMIN, UserRole.BRANCH_ADMIN)
-  @ApiBearerAuth()
-  @ApiOperation({
-    summary: 'Create a new product',
-    description: 'Only ADMIN and BRANCH_ADMIN can create a product.',
-  })
-  @ApiBody({ type: CreateProductDTO })
-  @ApiCreatedResponse({
-    description: 'Product successfully created.',
-    type: Product,
-  })
-  @ApiUnauthorizedResponse({ description: 'User is not authorized.' })
-  async createProduct(
-    @Body() createProductDto: CreateProductDTO,
-  ): Promise<Product> {
-    const manufacturer = await this.productsServices.findManufacturer(
-      createProductDto.manufacturer,
-    );
-
-    const newProduct = await this.productsServices.createProduct(
-      createProductDto,
-      manufacturer,
-    );
-
-    if (createProductDto.imageUrls && createProductDto.imageUrls.length) {
-      await this.productsServices.createProductImage(
-        newProduct,
-        createProductDto.imageUrls,
-      );
-    }
-
-    if (createProductDto.categoryIds && createProductDto.categoryIds.length) {
-      const categories = await this.productsServices.findCategories(
-        createProductDto.categoryIds,
-      );
-
-      await this.productsServices.addCategoriesToProduct(
-        newProduct,
-        categories,
-      );
-    }
-
-    if (
-      createProductDto.presentations &&
-      createProductDto.presentations.length
-    ) {
-      const ids = createProductDto.presentations.map((p) => p.presentationId);
-      const presentations = await this.productsServices.findPresentations(ids);
-
-      await this.productsServices.addPresentationsToProduct(
-        newProduct,
-        presentations,
-        createProductDto.presentations,
-      );
-    }
-
-    return newProduct;
   }
 }
