+++ src/products/products.service.ts
@@ -1,12 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Product } from './entities/product.entity';
-import { In, Repository, Brackets, SelectQueryBuilder, IsNull } from 'typeorm';
+import { Repository, Brackets, SelectQueryBuilder } from 'typeorm';
 import { ProductPresentation } from './entities/product-presentation.entity';
-import {
-  CreateProductDTO,
-  CreateProductPresentationDTO,
-} from './dto/product.dto';
 import { Manufacturer } from './entities/manufacturer.entity';
 import { ProductImage } from './entities/product-image.entity';
 import { Presentation } from './entities/presentation.entity';
@@ -125,144 +121,5 @@ export class ProductsService {
       .getMany();
 
     return { products, total };
-  }
-
-  async findOne(id: string): Promise<Product> {
-    const product = await this.productRepository.findOne({
-      where: { id, deletedAt: IsNull() },
-      relations: ['images'],
-    });
-    if (!product) {
-      throw new NotFoundException(`Product #${id} not found`);
-    }
-    return product;
-  }
-
-  async findManufacturer(id: string): Promise<Manufacturer> {
-    const manufacturer = await this.manufacturerRepository.findOne({
-      where: { id },
-    });
-
-    if (!manufacturer) {
-      throw new NotFoundException('Manufacturer not found');
-    }
-
-    return manufacturer;
-  }
-
-  async findCategories(ids: string[]): Promise<Category[]> {
-    if (!ids || ids.length === 0) {
-      return [];
-    }
-
-    const categories = await this.categoryRepository.findBy({
-      id: In(ids),
-    });
-
-    if (categories.length !== ids.length) {
-      throw new NotFoundException('One or more categories not found');
-    }
-
-    return categories;
-  }
-
-  async createProduct(
-    createProductDto: CreateProductDTO,
-    manufacturer: Manufacturer,
-  ): Promise<Product> {
-    const newProduct = this.productRepository.create({
-      ...createProductDto,
-      manufacturer,
-    });
-
-    const savedProduct = await this.productRepository.save(newProduct);
-    return savedProduct;
-  }
-
-  async createProductImage(product: Product, images: string[]): Promise<void> {
-    const productImages = images.map((url) =>
-      this.productImageRepository.create({ url, product }),
-    );
-    await this.productImageRepository.save(productImages);
-  }
-
-  async addCategoriesToProduct(
-    product: Product,
-    categoriesToAdd: Category[],
-  ): Promise<void> {
-    if (categoriesToAdd.length === 0) {
-      return;
-    }
-
-    if (!product.categories) {
-      product.categories = [];
-    }
-
-    product.categories = [...product.categories, ...categoriesToAdd];
-
-    await this.productRepository.save(product);
-  }
-
-  async findPresentations(ids: string[]): Promise<Presentation[]> {
-    if (!ids || ids.length === 0) {
-      return [];
-    }
-
-    const presentations = await this.PresentationRepository.findBy({
-      id: In(ids),
-    });
-
-    if (presentations.length !== ids.length) {
-      throw new NotFoundException('One or more presentations not found');
-    }
-
-    return presentations;
-  }
-
-  async addPresentationsToProduct(
-    product: Product,
-    presentations: Presentation[],
-    productPresentationDTOs: CreateProductPresentationDTO[],
-  ): Promise<void> {
-    if (productPresentationDTOs.length === 0) {
-      return;
-    }
-
-    const productPresentations = productPresentationDTOs.map((dto) => {
-      const presentation = presentations.find(
-        (p) => p.id === dto.presentationId,
-      );
-
-      if (!presentation) {
-        throw new NotFoundException(
-          `Presentation with ID ${dto.presentationId} not found`,
-        );
-      }
-
-      return this.productPresentationRepository.create({
-        product,
-        presentation,
-        price: dto.price,
-      });
-    });
-
-    await this.productPresentationRepository.save(productPresentations);
-  }
-
-  async findProductImage(productId: string, imageId: string) {
-    return this.productImageRepository.findOne({
-      where: {
-        id: imageId,
-        product: { id: productId },
-      },
-    });
-  }
-
-  async updateProductImage(image: ProductImage): Promise<ProductImage> {
-    return this.productImageRepository.save(image);
-  }
-
-  async deleteProductImage(image: ProductImage): Promise<void> {
-    await this.productImageRepository.softRemove(image);
   }
 }
